requirements.txt
aiohttp==3.6.2
aiounittest==1.4.0
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.5.4
black==20.8b1
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
idna==2.8
idna-ssl==1.1.0
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.5.2
mypy-extensions==0.4.3
netifaces==0.10.9
pathspec==0.8.0
pycodestyle==2.6.0
pylint==2.6.0
python-engineio==3.13.2
python-socketio==4.6.0
regex==2020.7.14
requests==2.22.0
six==1.12.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4
urllib3==1.25.3
websocket-client==0.56.0
websockets==8.1
wrapt==1.12.1
yarl==1.3.0
